+++ src/debug.cpp
@@ -54,21 +54,21 @@ int disassembleInstruction(const Chunk& chunk, int offset) {
     uint8_t instruction = chunk.code[offset];
 
     switch (instruction) {
-        case OpCode::OP_CONSTANT:
+        case static_cast<uint8_t>(OpCode::OP_CONSTANT):
             return constantInstruction("OP_CONSTANT", chunk, offset);
-        case OpCode::OP_CONSTANT_LONG:
+        case static_cast<uint8_t>(OpCode::OP_CONSTANT_LONG):
             return constantLongInstruction("OP_CONSTANT_LONG", chunk, offset);
-        case OpCode::OP_ADD:
+        case static_cast<uint8_t>(OpCode::OP_ADD):
             return simpleInstruction("OP_ADD", offset);
-        case OpCode::OP_SUBTRACT:
+        case static_cast<uint8_t>(OpCode::OP_SUBTRACT):
             return simpleInstruction("OP_SUBTRACT", offset);
-        case OpCode::OP_MULTIPLY:
+        case static_cast<uint8_t>(OpCode::OP_MULTIPLY):
             return simpleInstruction("OP_MULTIPLY", offset);
-        case OpCode::OP_DIVIDE:
+        case static_cast<uint8_t>(OpCode::OP_DIVIDE):
             return simpleInstruction("OP_DIVIDE", offset);
-        case OpCode::OP_NEGATE:
+        case static_cast<uint8_t>(OpCode::OP_NEGATE):
             return simpleInstruction("OP_NEGATE", offset);
-        case OpCode::OP_RETURN:
+        case static_cast<uint8_t>(OpCode::OP_RETURN):
             return simpleInstruction("OP_RETURN", offset);
         default:
             std::cout << std::format("Unknown opcode {}\n", instruction);
